refactor: 컨트롤러 역할에 따라 세분화
