--- src/main/java/roomescape/controller/AdminController.java
@@ -1,13 +1,0 @@
-package roomescape.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class AdminController {
-
-    @GetMapping("/admin")
-    public String admin() {
-        return "/admin/index";
-    }
-}
