+++ src/test/java/roomescape/controller/PageControllerTest.java
@@ -1,0 +1,74 @@
+package roomescape.controller;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class PageControllerTest {
+
+    @LocalServerPort
+    int port;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+    }
+
+    @DisplayName("어드민 페이지를 호출 시 200으로 응답한다.")
+    @Test
+    void adminPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("어드민 예약 페이지 호출 시 200으로 응답한다.")
+    @Test
+    void reservationPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("어드민 테마 페이지 요청 시 200으로 응답한다.")
+    @Test
+    void themePageTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin/theme")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("어드민 시간 페이지 요청 시 200으로 응답한다.")
+    @Test
+    void timeTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin/time")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("유저 예약 페이지 요청 시 200으로 응답한다.")
+    @Test
+    void welcomePageTest() {
+        RestAssured.given().log().all()
+                .when().get("/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @DisplayName("인기 테마 페이지 요청 시 200으로 응답한다.")
+    @Test
+    void bestPageTest() {
+        RestAssured.given().log().all()
+                .when().get("/")
+                .then().log().all()
+                .statusCode(200);
+    }
+}
